refactor: rename mcp server feature for consistency
