+++ packages/mcp-server/src/features/core/index.ts
@@ -4,7 +4,7 @@ import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js"
 import { registerFetchTool as setupFetch } from "../fetch";
 import { setup as setupLocalRestApi } from "../local-rest-api";
 import { setup as setupPrompts } from "../prompts";
-import { setup as setupSmartSearch } from "../smart-connections";
+import { setup as setupSmartSearch } from "../smart-search";
 import { setup as setupTemplates } from "../templates";
 import { setup as setupSourceDocuments } from "../source-document";
 import {
