+++ packages/mcp-server/src/features/smart-search/index.ts
